FILE RESTRUCTURE Moved everything to folder cgcu, except index.
This is a fix to the cgcu.net domain problem:
cgcu.net/ does correctly map to index.
cgcu.net/contact.html does not map to the root of cgcu directory on the union server.
cgcu.net/cgcu/ does.
the cgcu.net maps to union.ic.ac.uk/guilds/ only.

So, TO UPLOAD TO THE SITE:
All links should be relative in the current file structure.
When uploading, copy all of the cgcu-website/cgcu/ into guilds/cgcu on the union server, and copy the index.html into the same folder (guilds/cgcu/)
The folder guilds/cgcu/cgcu/ should not exist on the union server.